requirements.txt
aniso8601==8.0.0
appdirs==1.4.4
astroid==2.4.2
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-RESTful==0.3.8
gunicorn==20.0.4
idna==2.10
isort==5.0.6
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
lazy-object-proxy==1.5.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.4
pipenv==2020.6.2
psycopg2-binary==2.8.5
pycparser==2.20
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.9
virtualenv==20.0.26
virtualenv-clone==0.5.4
Werkzeug==1.0.1
wrapt==1.12.1
